+++ src/main.js
@@ -1,6 +1,5 @@
-import getData from './getData.js';
 import makeFilter from './make-filter.js';
-import makeTripPoint from './make-trip-point.js';
+import {renderTripPoint, removeTripPoint} from './make-trip-point.js';
 
 
 const filters = document.querySelector(`.trip-filter`);
@@ -8,13 +7,6 @@ const filters = document.querySelector(`.trip-filter`);
 filters.insertAdjacentHTML(`beforeend`, makeFilter(`Everything`, true));
 filters.insertAdjacentHTML(`beforeend`, makeFilter(`Future`));
 filters.insertAdjacentHTML(`beforeend`, makeFilter(`Past`));
-
-const renderTripPoint = (amount, dist) => {
-  const points = new Array(amount)
-    .fill()
-    .map(() => makeTripPoint(getData()));
-  dist.insertAdjacentHTML(`beforeend`, points.join(``));
-};
 
 const tripPointsContainer = document.querySelector(`.trip-day__items`);
 renderTripPoint(7, tripPointsContainer);
@@ -27,14 +19,3 @@ filtersArray.forEach(function (element) {
     renderTripPoint(amountTripPoins, tripPointsContainer);
   });
 });
-
-const removeTripPoint = function () {
-  const tripPointsArray = document.querySelectorAll(`.trip-point`);
-  if (tripPointsArray.length !== 0) {
-    tripPointsArray.forEach(function (element) {
-      tripPointsContainer.removeChild(element);
-    });
-  }
-};
-
-
